Correct file name and fix some bugs. Correct systax error: path -> patch In fact the header of preloader file is not only EMMC_BOOT and UFS_BOOT but also COMBO_BOOT and others. So I modified the header validation in order to ignore the difference quickly.
